move porting functionalities to separate file
